backend/math_backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
Pillow==10.0.0  # For image fields if needed